DAS-2180: renames harmony_service_entry -> harmony_service
This makes more sense since entry is tied to Docker.
