+++ harmony_service/exceptions.py
